backend/video-analysis/requirements.txt
opencv-python
ultralytics
moviepy<2.0.0
openai
fastapi
uvicorn
python-multipart
python-dotenv
SpeechRecognition
